Initial version of the project
This is the initial version of the project as provided by the QA representative for the Data Returns service. It has been iterated previously but his was outside of source control.

The changes from that which was handed over are

- Everything moved to a root `features` folder
- The csv files that the tests rely on added to the project rather than being maintained separately
- Removal of user specific hard coded file locations
- Addition of a README, LICENCE and .gitignore file
- Removal of duplicate step that was causing all tests to fail

Currently some of the tests are failing due to additional duplication of steps, or some steps unable to locate UI elements, but the majority are working. We are pushing this as the initial version to capture as close as possible what was handed over, before making any further changes.