Set url endpoint to test using configuration file
A digital service generally has multiple environments and data returns is no different. This means the ability to run the test suite against different url endpoints is needed.

This is currently achieved by changing the code in `features/step_definitions/set_environment.js`. Now that the project is under source control this is not maintainable going forward.

Therefore this change reconfigures the project to use a [configuration file](https://chimp.readme.io/docs/command-line-options#section-using-a-configuration-file). The actual configuration used won't be checked in, but an example one will be.