Integrate the project with travis
As a first step to adding this project to the CI process for the Data Returns service, we are integrating it with Travis-CI.

Not only will getting it working on their give us some insites to what is needed to get it running on Jenkins, but it will also mean that changes to the tests going forward will be automatically checked to ensure they do pass before being committed.

This change covers adding the `.travis.yml` file, and some housekeeping on the `.gitignore` file.