Resolve issue with Cucumber-js v1.2.1
https://github.com/xolvio/chimp/issues/441
https://github.com/cucumber/cucumber-js/issues/610

The latest version of Chimp uses Cucumber-js v1.2.1 which appears to have an issue when any unexpected error is thrown (e.g. element not found).

It looks like cucumber-js have fixed this, but until Chimp updates to the lastest version am downgrading Chimp to a version that uses Cucumber-js v1.2.0